Đạo

+++ 1.DAL/Models/hoaDonChiTiet.cs
@@ -11,7 +11,7 @@ namespace _1.DAL.Models
         public int maHoaDon { get; set; }
         public int maSanPham { get; set; }
         public int Soluong { get; set; }
-        public float donGia { get; set; }
+        public decimal donGia { get; set; }
         public bool Trangthai { get; set; }
         public SanPham sanPham { get; set; }
         public hoaDon hoaDon { get; set; }
